Rename package name to use standard practice of lowercase and to have unique global name to avoid match conflict
